Rename DataStore to JavaDataStore
